+++ test/libsolidity/syntaxTests/inheritance/duplicated_ancestor_constructor_call.sol
@@ -1,4 +1,5 @@
 contract A { constructor(uint) public { } }
-contract B is A(2) { constructor() A(3) public {  } }
+contract B is A(2) { constructor() public {  } }
+contract C is B { constructor() A(3) public {  } }
 // ----
 // Warning: Duplicated super constructor calls are deprecated.